Move user interface to shell helper
Benefits:

  * More stability: restart on crash
  * The compositor does what it's better at
  * Use hardware layers
  * Render only what needs to be rendered

While QtWayland Compositor has the appropriate infrastructure for
hardware layers, there is currently no implementation for DRM.

In the future we will also be able to be clever and paint scene graph nodes
differently, depending on the corresponding surface and try to squeeze the
maximum efficiency out of the graphics chip.

Layer surface items now support being inside a container, anchors and margins
are applied to said container.

Layer surfaces that are keyboard interactive, are given the focus as soon
as their view is shown.

The client can now register shortcuts with the compositor.

Minor optimizations: instead of rendering a gray rectangle beneath the
surfaces area, we set the window color.
This saves creating an object and makes the rendering faster because setting
the window color is cheaper.

Background is no longer an hardware layer, because in present mode it gets
resized along with the surfaces area and it's not full screen anymore and
this would not be possible to implement with hardware layers, because the
surface would always be rendered below the surfaces area.

Previously some modal dialogs were using the dark theme, while other were
using the light theme.  Now all use the light theme for consistency.

Indicators are rendered into their own surface, this way the lock screen can
slide between panel and indicators as it was designed in the beginning.

Known issues:

  * Clicking on indicators doesn't open the drawer anymore, it will
    be implemented in another commit.
  * When a layer surface is closed, focus is not given back to the
    window previously focused.

Closes: #184